fix: resolve 8 runtime bugs from code review
- zHandlers.ino: nmea[100] -> nmea[200] (buffer overflow on full GGA sentence ~141 bytes)
- zADS1115.cpp: split (Wire1.read()<<8)|Wire1.read() into hi/lo vars in getConversion() and isConversionDone() (unspecified eval order -> possible byte swap)
- AIO_Keya_WasKeyaFiltre.ino: int8_t -> int16_t for KeyaCurrentSensorReading (EMA converges to ~5120, overflows signed byte)
- FXUtil.cpp: guard line[nchar-1]=0 against nchar==0 (writes to line[-1] on empty serial read)
- Autosteer.ino: EEPROM.get(70,aogConfig) -> EEPROM.put on first boot (was loading garbage from virgin EEPROM)
- Autosteer.ino: remove wasOffset==0 block (fired on every PGN 252 from AOG default config, defeating auto-zero)
- SectionControl.ino + Autosteer.ino: add len param to updatePinMapping(), guard len<29 (OOB read on short UDP packet)
- zAutoZeroMenu.ino: fix reset-defaults literal (speedMin 1.0->2.5, gpsHdgMax 1.0->0.3)

+++ AIO_Keya_WasKeyaFiltre.ino
@@ -23,7 +23,7 @@ Stream* NmeaOutputSerial = &Serial2;   //NMEA out
 const int32_t baudAOG = 115200; 
 const int32_t baudGPS = 460800;
 
-int8_t KeyaCurrentSensorReading = 0;
+int16_t KeyaCurrentSensorReading = 0;
 
 #define ImuWire Wire        //SCL=19:A5 SDA=18:A4
 #define RAD_TO_DEG_X_10 572.95779513082320876798154814105
